requirements.txt
streamlit
openai
pinecone-client
transformers
gtts
speechrecognition
asyncio
numpy
pandas
torch
torchaudio
torchvision
tiktoken
huggingface-hub
matplotlib
seaborn
plotly
spacy
nltk
streamlit-option-menu
streamlit-aggrid
python-dotenv
sounddevice
soundfile
pyaudio